Nanobragg noise (#865)
Uupdate the exascale API to add_noise to a multipanel diffraction image.  Previously it only worked on monolithic detector construction.

* Move encapsulated_twodev class to the header file.

This way it can be utilized in more than one C++ translation units, each
of which implements noise.

* New nanoBragg::add_noise() const correctness.

Therefore instead of adding noise to the internal diffraction image array,
we add it to an array passed in as an argument.

* In the exafel API, Add noise to multipanel image.

Utilize the underlying nanoBragg::add_noise() function, but pass it
image arrays from one detector panel at a time.  Work in progress!  This
is a CPU implementation, which could eventually be migrated to GPU.
Takes 4 seconds for Jungfrau 16M.