Refactor `small_cell_index_detail` (#912)
* Float import

* Invert conditional so a large block of the function can be dedented

* Move calculation of orientation matrix and indexed reflections to its own function

This will allow e.g. dials.index to use this function without needing
the image integration code in small_cell_index_detail. In fact,
dials.index cannot do the integration steps because the experiment list
is loaded with check_format=False, and image data are not accessible.

* Move raw_data access to where it is needed

* Copy the setup definitions back into small_cell_index_detail

* Fix return values

* bugfix

* bugfixes

`ok_to_integrate` and `s0` were only defined in `small_cell_index_lattice_detail`.

In fact, `ok_to_integrate` is unnecessary, because it is the same as checking
whether `ori is not None`.