Automate the energy scan calibration using the FEE spectrometer
The process conducted at the start of every LCLS beam time using scripts not under version control, manual/visual interpolation of plots, and trend fitting in Excel is fully automated in python with this commit. This is only ever used at LCLS but seems more generally applicable, so only the parts unique to LCLS are in the xfel namespace and the rest is in serialtbx. energy_scan_notch_finder.py doesn't rely on any cctbx packages and can be used without a cctbx.xfel installation. fee_calibration.py uses cctbx only for phil parsing and exception handling and could be easily converted for users of other packages at LCLS or for lightweight incorporation into the elog auto-generated plots.