Bugfix: convert after polyfit
In the energy scan notch finder, a best fit line is found for a series of known energies at measured pixel positions on the FEE detector. The best fit line is determined with polyfit, and the method's __call__ function works as expected, but the reported coefficients are not in the expected frame of reference until the convert method is called. Since we will use the raw coefficients elsewhere, we must call this method.