MMTBX: remove comparison of int with NoneType (#969)
* Do not compare NoneType to int

If twin_results_interpretation is instantiated with
translational_pseudo_symmetry is None, then self.patterson_height is
also None.  Therefore do not try to compare self.patterson_height to an
integer unless we know it has a numeric value.

* Again, do not compare NoneType to int

self.low_resolution_completeness_overall is initially given the value
None and only conditionally given a numeric value. We should not compare
self.low_resolution_completeness_overall to a numeric value if it is
None.