Add wavelengths to the run stats plot
Involves a change to the DB schema.  Before it used to be that wavelengths weren't logged if the images didn't index, and indeed an entry in the imageset table wasn't recorded.  Now, if the image didn't index, it's logged in the imageset table, just without a crystal, so that wavelength can also be logged.  Therefore the schema needs to loosen in to make the primary key for the experiment table not require a crystal model.

A consequence is that we now need to do a separate query for images that did or didn't not index, selecing on those without an imageset. Might revisit later to see if they can be combined into one query.