Spring 2025 XFEL GUI updates (#1065)
* Add options to load_cached_settings: scope=None, extract=True

Allows control of the scope and whether it is extracted

* Initialize debug file handle for small cell process and frame_logger

* For the notch finder, allow searching for minima or maxima

Due to a change in how the energy is calibrated. Before it was searching for notches, now it's searching for peaks from a monochromator. Might go back to a notch later.

* Add repr to _run class

* Add mpi_abort_on_exception to ensemble_refine_pipeline

* Better tag support for xfel.plot_uc_cloud plus use cached params

* xfel.print_run_status updates

- Use cached params
- Switch to intersection instead of union for tags

* More tasks for Phenix jobs

* Add completeness to merging stats

Includes 90% completeness cutoff resolution

* Add Vanessa Oklejas (@voklejas2) to acknowledgments

* Better way to find images in runstats

Before clicking on a dot only searched in the x direction for an image.  Now search in the x and y direction to find the nearest image.  Much easier to use now.  Also added for resolution quality plot.

Note, use picture space instead of data space for the search to remove problems of scale between the x and y axes (important for both n spots (y) when greatly different than n images (x), and for resolution (y) vs n images (x)).

* Job submission updates

- Rename output logs from .out to .log
- Reduce chunk size for ensemble refinement to 64
- Add string replacement to the command for Phenix jobs, not just the phil params
- Check evaluate_anom for Phenix jobs as a not multiprocessed job
- Extra debugging for new dataset versions

* Add wavelengths to the run stats plot

Involves a change to the DB schema.  Before it used to be that wavelengths weren't logged if the images didn't index, and indeed an entry in the imageset table wasn't recorded.  Now, if the image didn't index, it's logged in the imageset table, just without a crystal, so that wavelength can also be logged.  Therefore the schema needs to loosen in to make the primary key for the experiment table not require a crystal model.

A consequence is that we now need to do a separate query for images that did or didn't not index, selecting on those without an imageset. Might revisit later to see if they can be combined into one query.

* Better error checking for cctbx.xfel.weather

* Update syntax exceptions

Co-authored-by: Daniel Paley <dwpaley@lbl.gov>