Add pre_split option to time varying refinement
If True, load the input data and chunk it before submitting refinement jobs based on the total number of input lattices. If False, chunk the data by counting the number of expt/refl files.

Downside of pre_split is that it takes longer to submit jobs since the data has to be loaded and sorted up front by the XFEL GUI.  Upsdie is it uses sort_by_imageset_path_and_image_index so that the output is reproducible.