Predict more spots on stills, gather more signal.
For stills integration, after the time-dependent ensemble refinement, we
reintegrate the data in the cctbx.xfel.mpi_integrate program.  Here,
allow the option of expanding the Nave-parameter envelope that is used
to predict spots, in order to capture more signal. Dramatically increase
the overall multiplicity, possibly by a factor of 1.5. Use this phil
parameter to activate this feature:
integration.recruitment.expand_nave_parameters=True